backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dateutil==2.8.2
